[Refactor] MyButton 수정
