+++ HoloCure/Assets/1_Scripts/UI/Common.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c68f7ad0a19bf18409331ea21c4f84c9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
